+++ Control/Joint/Liftable.hs
@@ -1,0 +1,4 @@
+module Control.Joint.Liftable where
+
+class Liftable (eff :: * -> *) (schema :: * -> *) where
+	lift :: eff a -> schema a