+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 
 # 0.1.3
 * Define `unite` method to convert composition to transformer
+* Define `Modulator` class to map function over inner effects